fix(capped): cast EF write-back to momentum dtype (bf16 crash)
The winners-only refactor changed the EF factor from the 0-dim ef_decay
scalar to a dimensioned [N, k, 1] fp32 tensor. PyTorch promotion treats
these differently: bf16 * 0-dim fp32 stays bf16, but bf16 * dimensioned
fp32 promotes to fp32, and the in-place scatter_ into bf16 momentum then
raises "Expected self.dtype to be equal to src.dtype". Only fires with
bf16 optimizer state (real mixed-precision training); every existing test
used fp32 params, so the suite stayed green.

Fix per review: multiply in fp32, round once into M's dtype (a no-op on
all fp32 paths). Add a bf16-param smoke test over local/global_capped x
Dion2/NorDion2 so the promotion path can't regress. 25/25 on 2x/4x H100.

Reported-by: JohnLangford (1b NorDion2 bf16 run)

+++ dion/dion2.py
@@ -683,9 +683,13 @@ def dion2_pre_orthogonalize(
     # Apply error feedback decay to selected slices in the original M tensors
     # (capped scope: winners only). We reuse the already-gathered slices and
     # write them back (scaled) using scatter_, which places values into
-    # positions specified by the index tensor.
+    # positions specified by the index tensor. The .to(dtype) matters for bf16
+    # momentum: the capped ef_factor is a *dimensioned* fp32 tensor, so
+    # bf16 * fp32 promotes to fp32 (unlike the 0-dim ef_decay scalar, which
+    # loses promotion to a dimensioned bf16), and scatter_ requires src dtype
+    # to match M exactly. Multiply in fp32, round once into M's dtype.
     indices_list = list(indices.unbind(dim=0))
-    ef_src_list = list((selected_stacked * ef_factor).unbind(dim=0))
+    ef_src_list = list((selected_stacked * ef_factor).to(dtype).unbind(dim=0))
     for m, idx, ef_src in zip(M, indices_list, ef_src_list):
         if select_dim == -2:
             idx_exp = idx.unsqueeze(-1).expand(*idx.shape, m.size(-1))
